+++ MaxBot/Objects/Stat.cs
@@ -4,5 +4,5 @@ namespace MaxBot.Objects;
 public class Stat
 {
     [JsonPropertyName("views")]
-    public int? Views { get; set; }
+    public long? Views { get; set; }
 }
